Move the shortcode to a separate file in "includes" folder
